Add files via upload
Pressing shift while using the EQ gain knobs will control the gain of every stem separate.

While shift is pressed and the pads are on the "cue" option, the stems can turned on or off using the pads. To delete a cue point, you have to press cue again so it blinks and then press shift and the cue point you want to delete.

The four buttons around the library scroll knob selects the channel filterFX.

The crossfader curve selector knob is disabled.